test(matching): extract scoring helpers and add bidirectional language compatibility tests
Scenario: Strict bidirectional match scores highest
Scenario: One-directional match scores lower
Scenario: No language overlap scores zero
Task: #113
