+++ packages/api/package.json
@@ -9,7 +9,9 @@
       "default": "./src/*.ts"
     }
   },
-  "scripts": {},
+  "scripts": {
+    "test": "bun test"
+  },
   "dependencies": {
     "@sip-and-speak/auth": "workspace:*",
     "@sip-and-speak/db": "workspace:*",
